+++ examples/spring-streams-example/src/test/kotlin/com/stove/spring/streams/example/e2e/TestHelper.kt
@@ -5,8 +5,34 @@ import io.confluent.kafka.schemaregistry.testutil.MockSchemaRegistry
 import io.confluent.kafka.serializers.AbstractKafkaSchemaSerDeConfig.SCHEMA_REGISTRY_URL_CONFIG
 import io.confluent.kafka.streams.serdes.protobuf.KafkaProtobufSerde
 import okio.ByteString.Companion.toByteString
+import org.apache.kafka.common.serialization.*
 import org.slf4j.LoggerFactory
 import java.util.*
+
+class StoveKafkaValueSerializer<T : Any> : Serializer<T> {
+  private val protobufSerde: KafkaProtobufSerde<Message> = TestHelper().createConfiguredSerdeForRecordValues()
+
+  override fun serialize(
+    topic: String,
+    data: T
+  ): ByteArray {
+    return when (data) {
+      is ByteArray -> data
+      else -> protobufSerde.serializer().serialize(topic, data as Message)
+    }
+  }
+}
+
+class StoveKafkaValueDeserializer<T : Any> : Deserializer<Message> {
+  private val protobufSerde: KafkaProtobufSerde<Message> = TestHelper().createConfiguredSerdeForRecordValues()
+
+  override fun deserialize(
+    topic: String,
+    data: ByteArray
+  ): Message {
+    return protobufSerde.deserializer().deserialize(topic, data)
+  }
+}
 
 class TestHelper {
   companion object {
